chore(payments-server): move tooltip to directory
- refs #3718
